Refactor: move methods to new class CurrencyHelper
