+++ monkeys/monkey_budget/money/__init__.py
@@ -1,3 +1,3 @@
 from .monetary import Monetary
-from .currency import Currency
+from .currency import Currency, CurrencyHelper
 from . import currencies
