+++ monkeys/monkey_budget/money/monetary.py
@@ -109,15 +109,3 @@ class Monetary:
             raise AttributeError("The currencies does not match")
         else:
             return True
-
-    @staticmethod
-    def get_currency_by_its_code(code: str = None) -> Currency:
-        try:
-            currency_dict = getattr(currencies, code)
-            return currency_dict
-        except AttributeError:
-            print(f"No such a currency with code {code}")
-
-    @staticmethod
-    def get_currencies_set() -> list[tuple[str, str]]:
-        return [(currency["code"], currency["code"]) for currency in currencies.currencies_tuple]
